fix(EntityModel): denote obsolete property in Account model
Note IsComplianceTradable as an obsolete property, to be replaced with an account
attribute.
