+++ NjordinSight.DAL/DataTransfer/Deprecated/AccountCompositeDto.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace NjordinSight.DataTransfer
+namespace NjordinSight.DataTransfer.Deprecated
 {
     public class AccountCompositeDto : AccountObjectDto
     {
