+++ NjordinSight.DAL/DataTransfer/Deprecated/CountryAttributeGrouping.cs
@@ -4,38 +4,37 @@ using NjordinSight.EntityModel;
 using NjordinSight.EntityModel.Annotations;
 using NjordinSight.DataTransfer.Generic;
 
-namespace NjordinSight.DataTransfer
+namespace NjordinSight.DataTransfer.Deprecated
 {
-    [ModelAttributeSupport(
-            SupportedScopes = ModelAttributeScopeCode.Country | ModelAttributeScopeCode.Security)]
+    [ModelAttributeSupport(SupportedScopes = ModelAttributeScopeCode.Country)]
     /// <summary>
-    /// Represents a collection of <see cref="SecurityAttributeMemberEntry"/> instances with the same 
-    /// <see cref="Security" />, <see cref="ModelAttribute"/>, and effective date.
+    /// Represents a collection of <see cref="CountryAttributeMemberEntry"/> instances with the same 
+    /// <see cref="Country" />, <see cref="ModelAttribute"/>, and effective date.
     /// </summary>
-    public class SecurityAttributeGrouping :
-        AttributeEntryWeightedGrouping<Security, SecurityAttributeMemberEntry>
+    public class CountryAttributeGrouping :
+        AttributeEntryWeightedGrouping<Country, CountryAttributeMemberEntry>
     {
-        public SecurityAttributeGrouping(
-            Security parentEntity,
+        public CountryAttributeGrouping(
+            Country parentEntity,
             ModelAttribute modelAttribute,
             DateTime effectiveDate) : base(parentEntity, modelAttribute, effectiveDate)
         {
         }
 
-        protected override Func<Security, ICollection<SecurityAttributeMemberEntry>>
-            ParentEntryMemberSelector => x => x.SecurityAttributeMemberEntries;
+        protected override Func<Country, ICollection<CountryAttributeMemberEntry>>
+            ParentEntryMemberSelector => x => x.CountryAttributeMemberEntries;
 
-        protected override Func<SecurityAttributeMemberEntry, bool> EntrySelector => x =>
+        protected override Func<CountryAttributeMemberEntry, bool> EntrySelector => x =>
             (x.AttributeMember is null || x.AttributeMember.AttributeId == ParentAttribute.AttributeId)
                 && x.EffectiveDate == EffectiveDate;
 
-        protected override Func<SecurityAttributeMemberEntry, decimal> WeightSelector => x => x.Weight;
+        protected override Func<CountryAttributeMemberEntry, decimal> WeightSelector => x => x.Weight;
 
-        public override SecurityAttributeMemberEntry AddNewEntry()
+        public override CountryAttributeMemberEntry AddNewEntry()
         {
-            SecurityAttributeMemberEntry newEntry = new()
+            CountryAttributeMemberEntry newEntry = new()
             {
-                SecurityId = ParentObject.SecurityId,
+                CountryId = ParentObject.CountryId,
                 EffectiveDate = EffectiveDate,
                 Weight = default,
                 AttributeMemberId = default,
@@ -53,7 +52,7 @@ namespace NjordinSight.DataTransfer
         }
 
         protected override bool UpdateEntryEffectiveDate(
-            SecurityAttributeMemberEntry entry, DateTime effectiveDate)
+            CountryAttributeMemberEntry entry, DateTime effectiveDate)
         {
             entry.EffectiveDate = effectiveDate;
             return entry.EffectiveDate == effectiveDate;
