+++ NjordinSight.DAL/DataTransfer/Deprecated/PaginationData.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NjordinSight.DataTransfer
+namespace NjordinSight.DataTransfer.Deprecated
 {
     /// <summary>
     /// Represents a small collection of metadata describing the pagination of query
