+++ NjordinSight.DAL/DataTransfer/Deprecated/_ViewModelExtension.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace NjordinSight.DataTransfer
+namespace NjordinSight.DataTransfer.Deprecated
 {
     /// <summary>
     /// Extensions methods used by the Model.ViewModel namespace.
