+++ NjordinSight.DAL/DataTransfer/Generic/AttributeEntryWeightedCollection.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using NjordinSight.DataTransfer.Deprecated;
 using NjordinSight.EntityModel;
 
 namespace NjordinSight.DataTransfer.Generic
