+++ NjordinSight.DAL/NjordinSight.DAL.csproj
@@ -9,6 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
     <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
   </ItemGroup>
