+++ NjordinSight.Web/_Imports.razor
@@ -49,7 +49,7 @@
 @using NjordinSight.BusinessLogic.Functions
 @using NjordinSight.BusinessLogic.MarketFeed
 
-@using NjordinSight.DataTransfer
+@using NjordinSight.DataTransfer.Deprecated
 @using NjordinSight.DataTransfer.Generic
 @using NjordinSight.DataTransfer.MethodParams
 @using NjordinSight.EntityModel
